Move metrics to metrics package
